EDELIVERY-4054 Existing password is being accepted in change password
Add support for users editing their details by clicking the icon in the
right top menu when they are logged in.
Fix issues with the tooltip position in the main application component.
Ensure the identifier gets returned when the user logs in.
Add a secure REST endpoint to allow the users to update their own
details.